S20/requirements.txt
torch
torchvision
pillow
numpy
pandas
transformers
diffusers
scipy
accelerate